backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
decouple==0.0.7
distlib==0.3.6
Django==4.2.4
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
filelock==3.9.1
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
multidict==6.0.4
openai==0.28.1
Pillow==10.0.0
platformdirs==3.1.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.21.0
yarl==1.9.2
